HomeScreen.js swapped out w my code, some more changes to dependcies are made, overall the appication is not working
