+++ app.json
@@ -2,6 +2,7 @@
   "expo": {
     "name": "MotorVision",
     "slug": "MotorVision",
+    "entryPoint": "./App.js",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
