+++ package.json
@@ -27,7 +27,8 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-vector-icons": "^10.2.0",
-    "react-native-web": "~0.19.13"
+    "react-native-web": "~0.19.13",
+    "react-native-svg": "^12.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
